Roy/requirements.txt
numpy
geopandas
geopy
matplotlib
global_land_mark
streetview
selenium
rich
Pillow